+++ Runtime/HeatmapChart.cs
@@ -72,7 +72,7 @@ namespace XCharts
                     var rate = Random.Range(0, 101);
                     if (rate > 70) value = Random.Range(8f, 10f);
                     else value = Random.Range(1f, 8f);
-                    var list = new List<float> { i, j, value };
+                    var list = new List<double> { i, j, value };
                     AddData(0, list);
                 }
             }
